updated: mahasiswa can update progress academic

+++ src/controllers/mahasiswaController.js
@@ -158,11 +158,11 @@ const getProfileMahasiswaController = async (req, res) => {
 
 // !!! CHECK APAKAH SEMESTER YANG AKAN DIINPUT VALID
 const entryDataIrsController = async (req, res) => {
-  const { nim, semester, status, jumlahSks } = req.body;
+  const { nim, semester, status, jumlahSks, oldSemester } = req.body;
   const dokumen = req.file;
 
   // check null input
-  if (!nim || !semester || !status || !jumlahSks || !dokumen) {
+  if (!nim || !semester || !status || !jumlahSks) {
     if (dokumen) {
       fs.unlink(`public/documents/${dokumen.originalname}`, (err) => {
         if (err) throw err;
@@ -175,9 +175,11 @@ const entryDataIrsController = async (req, res) => {
 
   // Check nim
   if (nim != req.id) {
-    fs.unlink(`public/documents/${dokumen.originalname}`, (err) => {
-      if (err) throw err;
-    });
+    if (dokumen) {
+      fs.unlink(`public/documents/${dokumen.originalname}`, (err) => {
+        if (err) throw err;
+      });
+    }
     return res.status(403).json({
       message: "NIM berbeda dari data login. Entry tidak dapat dilakukan",
     });
@@ -188,9 +190,11 @@ const entryDataIrsController = async (req, res) => {
   // Check status
   const statusIRS = ["Aktif", "Cuti"];
   if (!statusIRS.includes(status)) {
-    fs.unlink(`public/documents/${dokumen.originalname}`, (err) => {
-      if (err) throw err;
-    });
+    if (dokumen) {
+      fs.unlink(`public/documents/${dokumen.originalname}`, (err) => {
+        if (err) throw err;
+      });
+    }
     return res.status(400).json({
       message: "Status IRS tidak valid",
     });
@@ -198,23 +202,28 @@ const entryDataIrsController = async (req, res) => {
 
   // Check jumlah sks
   if (jumlahSks < 0 || jumlahSks > 24) {
-    fs.unlink(`public/documents/${dokumen.originalname}`, (err) => {
-      if (err) throw err;
-    });
+    if (dokumen) {
+      fs.unlink(`public/documents/${dokumen.originalname}`, (err) => {
+        if (err) throw err;
+      });
+    }
     return res.status(400).json({
       message: "Jumlah SKS tidak valid",
     });
   }
 
   // Check file
-  if (path.extname(dokumen.originalname) !== ".pdf") {
-    // Errornya ngga json?
-    fs.unlink(`public/documents/${dokumen.originalname}`, (err) => {
-      if (err) throw err;
-    });
-    return res.status(400).json({
-      message: "Format dokumen harus pdf",
-    });
+  if (dokumen) {
+    if (path.extname(dokumen.originalname) !== ".pdf") {
+      // Errornya ngga json?
+
+      fs.unlink(`public/documents/${dokumen.originalname}`, (err) => {
+        if (err) throw err;
+      });
+      return res.status(400).json({
+        message: "Format dokumen harus pdf",
+      });
+    }
   }
 
   try {
@@ -224,6 +233,7 @@ const entryDataIrsController = async (req, res) => {
       status,
       jumlahSks,
       dokumen,
+      oldSemester,
     };
 
     const result = await entryDataIrs(data);
@@ -247,6 +257,7 @@ const entryDataKhsController = async (req, res) => {
     ips,
     jumlahSksKumulatif,
     ipk,
+    oldSemester,
   } = req.body;
   const dokumen = req.file;
 
@@ -258,8 +269,7 @@ const entryDataKhsController = async (req, res) => {
     !jumlahSksSemester ||
     !ips ||
     !jumlahSksKumulatif ||
-    !ipk ||
-    !dokumen
+    !ipk
   ) {
     if (dokumen) {
       fs.unlink(`public/documents/${dokumen.originalname}`, (err) => {
@@ -273,9 +283,11 @@ const entryDataKhsController = async (req, res) => {
 
   // Check nim
   if (nim != req.id) {
-    fs.unlink(`public/documents/${dokumen.originalname}`, (err) => {
-      if (err) throw err;
-    });
+    if (dokumen) {
+      fs.unlink(`public/documents/${dokumen.originalname}`, (err) => {
+        if (err) throw err;
+      });
+    }
     return res.status(403).json({
       message: "NIM berbeda dari data login. Entry tidak dapat dilakukan",
     });
@@ -285,9 +297,11 @@ const entryDataKhsController = async (req, res) => {
 
   // Check jumlah sks
   if (jumlahSksSemester < 0 || jumlahSksSemester > 24) {
-    fs.unlink(`public/documents/${dokumen.originalname}`, (err) => {
-      if (err) throw err;
-    });
+    if (dokumen) {
+      fs.unlink(`public/documents/${dokumen.originalname}`, (err) => {
+        if (err) throw err;
+      });
+    }
     return res.status(400).json({
       message: "Jumlah SKS tidak valid",
     });
@@ -295,9 +309,11 @@ const entryDataKhsController = async (req, res) => {
 
   // Check IPS
   if (parseFloat(ips) < 0 || parseFloat(ips) > 4) {
-    fs.unlink(`public/documents/${dokumen.originalname}`, (err) => {
-      if (err) throw err;
-    });
+    if (dokumen) {
+      fs.unlink(`public/documents/${dokumen.originalname}`, (err) => {
+        if (err) throw err;
+      });
+    }
     return res.status(400).json({
       message: "IPS tidak valid",
     });
@@ -307,16 +323,19 @@ const entryDataKhsController = async (req, res) => {
 
   // Check IPK
   if (parseFloat(ipk) < 0 || parseFloat(ipk) > 4) {
-    fs.unlink(`public/documents/${dokumen.originalname}`, (err) => {
-      if (err) throw err;
-    });
+    if (dokumen) {
+      fs.unlink(`public/documents/${dokumen.originalname}`, (err) => {
+        if (err) throw err;
+      });
+    }
     return res.status(400).json({
       message: "IPK tidak valid",
     });
   }
 
   // Check dokumen
-  if (path.extname(dokumen.originalname) !== ".pdf") {
+
+  if (dokumen && path.extname(dokumen.originalname) !== ".pdf") {
     fs.unlink(`public/documents/${dokumen.originalname}`, (err) => {
       if (err) throw err;
     });
@@ -335,6 +354,7 @@ const entryDataKhsController = async (req, res) => {
       jumlahSksKumulatif,
       ipk,
       dokumen,
+      oldSemester,
     };
 
     const result = await entryDataKhs(data);
@@ -349,11 +369,11 @@ const entryDataKhsController = async (req, res) => {
 };
 
 const entryDataPklController = async (req, res) => {
-  const { nim, semester, nilai } = req.body;
+  const { nim, semester, nilai, oldSemester } = req.body;
   const dokumen = req.file;
 
   // check null input
-  if (!nim || !semester || !nilai || !dokumen) {
+  if (!nim || !semester || !nilai) {
     if (dokumen) {
       fs.unlink(`public/documents/${dokumen.originalname}`, (err) => {
         if (err) throw err;
@@ -365,9 +385,11 @@ const entryDataPklController = async (req, res) => {
   }
   // Check nim
   if (nim != req.id) {
-    fs.unlink(`public/documents/${dokumen.originalname}`, (err) => {
-      if (err) throw err;
-    });
+    if (dokumen) {
+      fs.unlink(`public/documents/${dokumen.originalname}`, (err) => {
+        if (err) throw err;
+      });
+    }
     return res.status(403).json({
       message: "NIM berbeda dari data login. Entry tidak dapat dilakukan",
     });
@@ -375,9 +397,11 @@ const entryDataPklController = async (req, res) => {
 
   // Check semester
   if (!validateSemester(nim, semester)) {
-    fs.unlink(`public/documents/${dokumen.originalname}`, (err) => {
-      if (err) throw err;
-    });
+    if (dokumen) {
+      fs.unlink(`public/documents/${dokumen.originalname}`, (err) => {
+        if (err) throw err;
+      });
+    }
     return res.status(400).json({
       message: "Semester tidak valid",
     });
@@ -385,10 +409,12 @@ const entryDataPklController = async (req, res) => {
 
   // TODO-VALIDATE: validasi nilai PKL
 
-  if (path.extname(dokumen.originalname) !== ".pdf") {
-    fs.unlink(`public/documents/${dokumen.originalname}`, (err) => {
-      if (err) throw err;
-    });
+  if (dokumen && path.extname(dokumen.originalname) !== ".pdf") {
+    if (dokumen) {
+      fs.unlink(`public/documents/${dokumen.originalname}`, (err) => {
+        if (err) throw err;
+      });
+    }
     return res.status(400).json({
       message: "Format dokumen harus pdf",
     });
@@ -400,6 +426,7 @@ const entryDataPklController = async (req, res) => {
       semester,
       nilai,
       dokumen,
+      oldSemester,
     };
 
     const result = await entryDataPkl(data);
@@ -414,18 +441,12 @@ const entryDataPklController = async (req, res) => {
 };
 
 const entryDataSkripsiController = async (req, res) => {
-  const { nim, semester, nilai, tanggalLulusSidang, lamaStudi } = req.body;
+  const { nim, semester, nilai, tanggalLulusSidang, lamaStudi, oldSemester } =
+    req.body;
   const dokumen = req.file;
 
   // check null input
-  if (
-    !nim ||
-    !semester ||
-    !nilai ||
-    !tanggalLulusSidang ||
-    !lamaStudi ||
-    !dokumen
-  ) {
+  if (!nim || !semester || !nilai || !tanggalLulusSidang || !lamaStudi) {
     if (dokumen) {
       fs.unlink(`public/documents/${dokumen.originalname}`, (err) => {
         if (err) throw err;
@@ -438,9 +459,11 @@ const entryDataSkripsiController = async (req, res) => {
 
   // Check nim
   if (nim != req.id) {
-    fs.unlink(`public/documents/${dokumen.originalname}`, (err) => {
-      if (err) throw err;
-    });
+    if (dokumen) {
+      fs.unlink(`public/documents/${dokumen.originalname}`, (err) => {
+        if (err) throw err;
+      });
+    }
     return res.status(403).json({
       message: "NIM berbeda dari data login. Entry tidak dapat dilakukan",
     });
@@ -448,9 +471,11 @@ const entryDataSkripsiController = async (req, res) => {
 
   // Check semester
   if (!validateSemester(nim, semester)) {
-    fs.unlink(`public/documents/${dokumen.originalname}`, (err) => {
-      if (err) throw err;
-    });
+    if (dokumen) {
+      fs.unlink(`public/documents/${dokumen.originalname}`, (err) => {
+        if (err) throw err;
+      });
+    }
     return res.status(400).json({
       message: "Semester tidak valid",
     });
@@ -458,7 +483,7 @@ const entryDataSkripsiController = async (req, res) => {
 
   // TODO-VALIDATE: Check nilai skripsi, lama studi, dan tanggalLulusSidang
 
-  if (path.extname(dokumen.originalname) !== ".pdf") {
+  if (dokumen && path.extname(dokumen.originalname) !== ".pdf") {
     return res.status(400).json({
       message: "Format dokumen harus pdf",
     });
@@ -472,6 +497,7 @@ const entryDataSkripsiController = async (req, res) => {
       tanggalLulusSidang,
       dokumen,
       lamaStudi,
+      oldSemester,
     };
 
     const result = await entryDataSkripsi(data);
